requirements.txt
Django==4.1.7
django-browser-reload==1.6.0
djangorestframework==3.14.0